adding graph routes
